detection_node/requirements.txt
rclpy
numpy
ultralytics
eclipse-zenoh
opencv-python-headless
